requirements.txt
matplotlib==3.10.3
numpy==1.24.3
opencv_python==4.11.0.86
pandas==2.3.1
Pillow==11.3.0
scikit_learn==1.7.1
seaborn==0.13.2
torch==2.1.0
torchvision==0.16.0
tqdm==4.67.1
